+++ cmd/root.go
@@ -1,0 +1,26 @@
+package cmd
+
+import (
+	"github.com/spf13/cobra"
+)
+
+var (
+	cmd *cobra.Command
+
+	asHtml bool
+
+	fromCategory string
+	fromBook     string
+)
+
+func init() {
+	cmd = &cobra.Command{
+		Use:   "today",
+		Short: "Read some scripture today",
+	}
+}
+
+func Execute() {
+	err := cmd.Execute()
+	cobra.CheckErr(err)
+}